+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @redocly/cli
 
+## 2.30.3
+
+### Patch Changes
+
+- Updated `@redocly/ajv` to `v8.18.1`
+- Updated @redocly/respect-core to v2.30.3.
+
+## 2.30.2
+
+### Patch Changes
+
+- Improved Redocly config validation: now the config checks for typos in built-in decorator names.
+- Updated @redocly/openapi-core to v2.30.2.
+
+## 2.30.1
+
+### Patch Changes
+
+- Increased the default fetch timeout used by the `push` command to better support slower uploads.
+- Updated @redocly/openapi-core to v2.30.1.
+
+## 2.30.0
+
+### Patch Changes
+
+- Updated @redocly/openapi-core to v2.30.0.
+
 ## 2.29.2
 
 ### Patch Changes